--- a/rpcs3/Emu/Cell/SPUThread.h
+++ b/rpcs3/Emu/Cell/SPUThread.h
@@ -815,7 +815,7 @@ public:
 	const u32 option; // sys_spu_thread_initialize option
 	const u32 lv2_id; // The actual id that is used by syscalls
 
-	atomic_t<spu_thread*> next_cpu{}; // LV2 thread queues' node link
+	spu_thread* next_cpu{}; // LV2 thread queues' node link
 
 	// Thread name
 	atomic_ptr<std::string> spu_tname;