@elad335 2022-07-21
The latest commit reduces the time a made-up test finishes by 4 times compared to master, all the test does is lock and unlock sys_mutex_t on multiple threads. Keep in mind this won't scale as nicely to games.