@solarmystic 2022-07-22
Performance remains roughly the same, with similar CPU utilization in the games tested below:-

**GOW 3** 
This PR - 99 FPS/72 1% FPS/60 0.1% FPS - 54% CPU Usage
![GOW3 12378](https://user-images.githubusercontent.com/4345150/180395215-a9950020-5633-48f8-a019-f50a4d893442.PNG)

Current master - 99 FPS/72 1% FPS/59 0.1% FPS - 54% CPU Usage
![GOW3 master](https://user-images.githubusercontent.com/4345150/180395233-b6a30973-6965-4028-97ca-784bee316390.PNG)

**Persona 5**
This PR - 80 FPS/66 FPS 1% FPS/58 0.1% FPS - 67% CPU Usage
![P5 12378](https://user-images.githubusercontent.com/4345150/180395291-b0f69512-9292-4415-a207-f54cddcd5131.PNG)

Current master  - 81 FPS/67 FPS 1% FPS/61 0.1% FPS - 67% CPU Usage
![P5 master](https://user-images.githubusercontent.com/4345150/180395314-bdb67708-1854-4a92-8cb4-3667c6659f55.PNG)

**Demon's Souls**
This PR - 81 FPS/58 1% FPS/52 0.1% FPS - 33% CPU usage
![DS 12378](https://user-images.githubusercontent.com/4345150/180395337-b701e6ab-37be-4fc3-bda0-2c1c652b4bab.PNG)

Current master - 81 FPS/58 1% FPS/51 0.1% FPS - 34% CPU usage
![DS master](https://user-images.githubusercontent.com/4345150/180395360-6acc34d8-964f-4e16-9242-bff5230117c1.PNG)