@elad335 2022-07-22
Same when increasing PPU threads count (debug tab)? Especially on low thread count cpus.