@cipherxof 2022-07-28
Performance seems to be about the same or slightly worse in MGS4.

Master:

![2](https://user-images.githubusercontent.com/5994581/181461717-02fa92b9-989b-4fa9-944f-377b3c51604b.png)

PR:

![2](https://user-images.githubusercontent.com/5994581/181461753-e228fcef-25a4-41ad-8ab1-3df0c2a24b3c.png)