@elad335 2022-07-28
@cipherxof Send log please.