@cipherxof 2022-07-28
Upon re-testing the results were much closer, within margin of error. I'm not sure what the difference was in the first test.

Master:
[RPCS3.log.gz](https://github.com/RPCS3/rpcs3/files/9207516/RPCS3.log.gz)

PR:
[RPCS3.log.gz](https://github.com/RPCS3/rpcs3/files/9207534/RPCS3.log.gz)