@elad335 2022-07-28
Please test performance in games with the new commit.