@cipherxof 2022-07-28
Saw no performance difference.

[RPCS3.log.gz](https://github.com/RPCS3/rpcs3/files/9214274/RPCS3.log.gz)

![2](https://user-images.githubusercontent.com/5994581/181639026-2f7c5416-5867-48aa-80f8-8ce4a079f7d6.png)