@AhayriSG 2022-07-29
Skyrim BLUS30778 - Bleak Falls Barrow first scene
PR [130c45d](https://github.com/RPCS3/rpcs3/commit/130c45d21fd50e3163036b3e11289eb72cd253c7)
![PR rpcs3-Intel Core i7-6700K-NVIDIA GeForce GTX 1070-16GB (2x8GB) 3000MTs-frametimes](https://user-images.githubusercontent.com/59465544/181763303-151aa326-a316-438e-90f7-61f4f3342b81.svg)
master [0.0.23-13981](https://github.com/RPCS3/rpcs3/pull/12419)
![master rpcs3-Intel Core i7-6700K-NVIDIA GeForce GTX 1070-16GB (2x8GB) 3000MTs-master (0 0 23-13981)-frametimes](https://user-images.githubusercontent.com/59465544/181763341-7f3ec551-c7c6-4c6a-ba17-e4b91ca74644.svg)

config;
[config_BLUS30778.zip](https://github.com/RPCS3/rpcs3/files/9219985/config_BLUS30778.zip)