@elad335 2022-08-04
Eliminated all (I hope) the performance bugs for low thread-count CPUs.