@AhayriSG 2022-08-04
Same config: https://github.com/RPCS3/rpcs3/pull/12378#issuecomment-1199244137
![image](https://user-images.githubusercontent.com/59465544/182837033-75fe49c2-a1ca-42d6-b05b-bf7b11ae1216.png)
![image](https://user-images.githubusercontent.com/59465544/182837134-dcbed45a-c657-4155-9b62-17d637a9578c.png)
![image](https://user-images.githubusercontent.com/59465544/182837160-6159519c-64d4-449e-9aba-0a4ddddbd1cf.png)
[json files for capframex.zip](https://github.com/RPCS3/rpcs3/files/9259265/json.files.for.capframex.zip)