@AhayriSG 2022-08-04
c2da589c
[json file for capframex.zip](https://github.com/RPCS3/rpcs3/files/9260027/json.file.for.capframex.zip)
![image](https://user-images.githubusercontent.com/59465544/182857530-b02058f1-3d31-438e-b872-72e86cb36fa7.png)
![image](https://user-images.githubusercontent.com/59465544/182857628-db9aeadf-0d60-4609-8e08-773bc280e97e.png)
![image](https://user-images.githubusercontent.com/59465544/182857688-4fc785fc-31f4-428a-989b-ddcc4b878b7d.png)
![image](https://user-images.githubusercontent.com/59465544/182857345-fdeaeaaf-fd12-4595-adb7-7e19159a159f.png)