@Darkhost1999 2022-08-04
[RPCS3.log](https://github.com/RPCS3/rpcs3/files/9260562/RPCS3.log)

Can't get in game anymore (AC Brotherhood) freezes at the end of spu cache waiting for sys module get ID prx