@Darkhost1999 2022-08-04
![image](https://user-images.githubusercontent.com/60384196/182886757-87375c67-d918-4223-b0f6-d04e4be50060.png)
[RPCS3.log](https://github.com/RPCS3/rpcs3/files/9261006/RPCS3.log)