@Darkhost1999 2022-08-04
Working same as master. Maybe better performance 
![image](https://user-images.githubusercontent.com/60384196/182910927-010aac06-80cc-484e-82a1-9a7ba20a50ea.png)
Remaining issues are not result of or the intent of this PR.
[RPCS3.log](https://github.com/RPCS3/rpcs3/files/9261862/RPCS3.log)