@elad335 2022-08-07
Latest commit greatly narrows the gap between 4 PPU threads and 2. Keep in mind that this is a stress test, 4 PPU threads should still be faster in compatible games if you have enough threads.