@Augusto7743 2022-08-08
Perhaps is possible see an better performance in low cpus (AMD FX).