@coolllman 2022-08-08
https://github.com/RPCS3/rpcs3/issues/12485