@NFSROADCHALLENGE 2022-08-10
gran turismo 6 bces01893 ver 1.5. no controller works, try xbox and ps4 controller no one works :(