@NFSROADCHALLENGE 2022-08-10
I have only amd ryzen 7 3700x and ryzen 5 3600xt :/

Hämta Outlook for Android<https://aka.ms/AAb9ysg>
________________________________
From: Augusto7743 ***@***.***>
Sent: Wednesday, August 10, 2022 8:23:51 AM
To: RPCS3/rpcs3 ***@***.***>
Cc: NFSROADCHALLENGE ***@***.***>; Comment ***@***.***>
Subject: Re: [RPCS3/rpcs3] LV2 Optimizations (PR #12378)


Have anyone tested in Intel TSX cpu with an OS disabled all security mitigations ? (spectre, meltdown and etc)

—
Reply to this email directly, view it on GitHub<https://github.com/RPCS3/rpcs3/pull/12378#issuecomment-1210215293>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/AZOXKYQQMIHYG7YZOCUDFILVYNDHPANCNFSM54G3BTCQ>.
You are receiving this because you commented.Message ID: ***@***.***>